refactor: restructure and dedup some code
