+++ packages/widget-playground/src/components/Widget/WidgetView.tsx
@@ -12,8 +12,6 @@ export function WidgetView() {
     drawerRef?.current?.toggleDrawer();
   }, []);
 
-  console.log('WidgetView', config);
-
   return (
     <WidgetViewContainer toggleDrawer={toggleDrawer}>
       <LiFiWidget
