+++ packages/widget-playground/src/types.ts
@@ -1,0 +1,3 @@
+export type Indexable = Record<string | number, any>;
+export type ObjectType = Record<string, any>;
+export type ArrayType = Array<any>;
