+++ packages/widget-playground/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './color';
 export * from './getValueFromPath';
-export * from './addValueToPath';
+export * from './addValue';
+export * from './deleteValue';
 export * from './shortenAddress';
